requirements.txt
flask
pillow
numpy
torch
realesrgan